khala-rvc/requirements.txt
torch
torchaudio
faiss-cpu
numpy
scipy
librosa
fairseq @ git+https://github.com/facebookresearch/fairseq.git
